Tidy up the Althea extensions to Babel
This commit aims to make Althea-specific changes to Babel more robust
and integrated with the implementation.

babeld.c:
* Switch the price to the uint32_t type
* Check the price and multiplier more strictly
* Add both of the new flags to the usage message

babeld.man:
* Add entries for the price and the quality multiplier

configuration.c:
* Make getuint() check for errno after the strtoul() call
* Add a config option for the quality multiplier

local.c:
* Explicitly list "full-path-rtt" in dumps (used to be just "rtt")

message.c:
* Correct variable names to explicitly mention the full path RTT

util.c:
* Remove the redundant parse_price function

xroute.c:
* Hardcode a 0 price value in add_xroute so that we don't bill our
neighbours and only for forwarding.

tests/multihop-gdb-rtt.sh:
* typo

+++ xroute.h
@@ -19,6 +19,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
+#include <stdint.h>
 
 struct xroute {
     unsigned char prefix[16];
@@ -26,7 +27,7 @@ struct xroute {
     unsigned char src_prefix[16];
     unsigned char src_plen;
     unsigned short metric;
-    unsigned short price;
+    uint32_t price;
     unsigned int ifindex;
     int proto;
 };
